{/vvgBankomatWdlPrepaid/src/main/java/vvgbankomat/izborIznosaDat.java}

+++ vvgBankomatWdlPrepaid/src/main/java/vvgbankomat/izborIznosaDat.java
@@ -11,23 +11,23 @@ public class izborIznosaDat implements java.io.Serializable
    static final long serialVersionUID = 1L;
 
    @org.kie.api.definition.type.Label("IzabranIznos")
-   private Double izabranIznos;
+   private Integer izabranIznos;
 
    public izborIznosaDat()
    {
    }
 
-   public java.lang.Double getIzabranIznos()
+   public java.lang.Integer getIzabranIznos()
    {
       return this.izabranIznos;
    }
 
-   public void setIzabranIznos(java.lang.Double izabranIznos)
+   public void setIzabranIznos(java.lang.Integer izabranIznos)
    {
       this.izabranIznos = izabranIznos;
    }
 
-   public izborIznosaDat(java.lang.Double izabranIznos)
+   public izborIznosaDat(java.lang.Integer izabranIznos)
    {
       this.izabranIznos = izabranIznos;
    }
